[:sparkles: feat] 카카오 소셜로그인 구현

+++ src/main/java/speedit/bookplate/utils/JwtService.java
@@ -21,7 +21,7 @@ public class JwtService {
     @param userIdx
     @return String
      */
-    public String createJwt(int userIdx){
+    public String createJwt(long userIdx){
         Date now = new Date();
         return Jwts.builder()
                 .setHeaderParam("type","jwt")
